% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_cn.r
\name{BUILD_cn_normal_XY}
\alias{BUILD_cn_normal_XY}
\title{Dataframe of information about chromosome bins}
\usage{
BUILD_cn_normal_XY(TABLE_CHROMOSOME_CN_INFO)
}
\arguments{
\item{TABLE_CHROMOSOME_CN_INFO}{A dataframe containing columns "Chromosome", "Bin_count", "Centromere_location". "Chromosome" refers to the chromosome number. "Bin_count" refers to the number of bins in each chromosome. "Centromere_location" refers to the approximate base pair position of the centromere of the chromosome.}
}
\description{
\code{BUILD_cn_normal_XY} returns a dataframe of information about each chromosome that composes a normal male genome (chromosome 1 to 22, each with 2 alleles A and B; and chromosomes X and Y, each with 1 allele A), including bin sizes and associated strand/allele.
}
\examples{

CN_bin_length <- 500000
vec_chromosome_name <- c(1:22, "X", "Y")
vec_chromosome_bp <- c(
       248956422, 242193529, 198295559, 190214555,
       181538259, 170805979, 159345973, 145138636,
       138394717, 133797422, 135086622, 133275309,
       114364328, 107043718, 101991189, 90338345,
       83257441, 80373285, 58617616, 64444167,
       46709983, 50818468, 156040895, 57227415)
vec_centromere_bp <- c(
       125, 93.3, 91, 50.4,
       48.4, 61, 59.9, 45.6,
       49, 40.2, 53.7, 35.8,
       17.9, 17.6, 19, 36.6,
       24, 17.2, 26.5, 27.5,
       13.2, 14.7, 60.6, 10.4) * 10^6
vec_bin_count <- ceiling(vec_chromosome_bp / CN_bin_length)
vec_centromere_location <- round(vec_centromere_bp / CN_bin_length)

TABLE_CHROMOSOME_CN_INFO <- data.frame(vec_chromosome_name, vec_bin_count, vec_centromere_location)
columns <- c("Chromosome", "Bin_count", "Centromere_location")
colnames(TABLE_CHROMOSOME_CN_INFO) <- columns

CN_matrix <- BUILD_cn_normal_XX(TABLE_CHROMOSOME_CN_INFO)

}
