% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_driver_library.r
\name{BUILD_driver_library}
\alias{BUILD_driver_library}
\title{Compute and add the selection rates for driver alleles based on the choice of selection model}
\usage{
BUILD_driver_library(
  model_variables = list(),
  table_arm_selection_rates = list(),
  table_gene_selection_rates = list(),
  vec_driver_genes = c(),
  vec_driver_role = c(),
  vec_chromosome = -1,
  vec_bin = -1,
  vec_driver_s = c(),
  vec_id = c(),
  vec_start = c(),
  vec_end = c(),
  vec_arm_s = c()
)
}
\arguments{
\item{model_variables}{The object returned by \code{BUILD_general_variables}.}

\item{table_arm_selection_rates}{A dataframe consisting of columns 'Arm_ID', 'Chromosome', 'Bin_start', 'Bin_end', 's_rate'. 'Arm_ID' refers to the name of each chromosome arm eg. 1p, 1q, etc. 'Chromosome' refers to the chromosome number. 'Bin_start' refers to the ???. 'Bin_end' refers to the ???. 's_rate' refers to the selective value of the chromosome arm as a float.}

\item{vec_driver_genes}{A list of driver genes. Only needed if \code{selection_model} is 'ancient'.}

\item{vec_driver_role}{A list corresponding to \code{vec_driver_genes} classifying each driver gene as a TSG or Oncogene. Only needed if \code{selection_model} is 'ancient'.}

\item{vec_chromosome}{???}

\item{vec_bin}{???}

\item{vec_driver_s}{A list corresponding to \code{vec_driver_genes} specifying the selection rate of the driver gene. Only needed if \code{selection_model} is 'ancient'.}

\item{vec_id}{???}

\item{vec_start}{???}

\item{vec_end}{???}

\item{vec_arm_s}{???}
}
\description{
\code{BUILD_driver_library} returns an updated version of \code{model_variables} that includes information about the chosen selection model and the resulting selection rates for different driver alleles based on whether they are classified as TSGs or Oncogenes.
}
\examples{

model_variables_base <- BUILD_driver_library(
   model_variables = model_variables_base,
   table_arm_selection_rates = data.frame(Arm_ID = arm_id, Chromosome = arm_chromosome, Bin_start = arm_start, Bin_end = arm_end, s_rate = arm_s))

}
