% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_general_variables.r
\name{BUILD_general_variables}
\alias{BUILD_general_variables}
\title{Produce simulations and outputs files as requested by user}
\usage{
BUILD_general_variables(
  model_name = "MODEL",
  cell_lifespan = 4,
  T_0 = list(0, "year"),
  T_end = list(100, "year"),
  T_tau_step = 3,
  Table_sample = data.frame(),
  table_population_dynamics = matrix(0, ncol = 2, nrow = 2),
  Population_end = Inf,
  Max_events = Inf,
  CN_bin_length = 5e+05,
  prob_CN_whole_genome_duplication = 0,
  prob_CN_missegregation = 0,
  prob_CN_chrom_arm_missegregation = 0,
  prob_CN_focal_amplification = 0,
  prob_CN_focal_deletion = 0,
  prob_CN_cnloh_interstitial = 0,
  prob_CN_cnloh_terminal = 0,
  prob_CN_focal_amplification_length = 0.1,
  prob_CN_focal_deletion_length = 0.1,
  prob_CN_cnloh_interstitial_length = 0.1,
  prob_CN_cnloh_terminal_length = 0.1,
  rate_driver = 0,
  rate_passenger = 0,
  bound_driver = 3,
  bound_average_ploidy = 5,
  bound_homozygosity = 0,
  SFS_totalsteps = 25,
  prob_coverage = 0.05,
  alpha_coverage = 0.7,
  lower_limit_cell_counts = 0,
  lower_limit_alt_counts = 3,
  lower_limit_tot_counts = 0,
  gc = data.frame(matrix(ncol = 5, nrow = 0)),
  gc_slope = 1.2,
  gc_int = 0,
  sigma1 = 0.1,
  num_reads = 1e+06
)
}
\arguments{
\item{model_name}{...}

\item{cell_lifespan}{...}

\item{T_0}{...}

\item{T_end}{...}

\item{T_tau_step}{...}

\item{Table_sample}{...}

\item{table_population_dynamics}{...}

\item{Population_end}{...}

\item{Max_events}{...}

\item{CN_bin_length}{...}

\item{prob_CN_whole_genome_duplication}{...}

\item{prob_CN_missegregation}{...}

\item{prob_CN_chrom_arm_missegregation}{...}

\item{prob_CN_focal_amplification}{...}

\item{prob_CN_focal_deletion}{...}

\item{prob_CN_cnloh_interstitial}{...}

\item{prob_CN_cnloh_terminal}{...}

\item{prob_CN_focal_amplification_length}{...}

\item{prob_CN_focal_deletion_length}{...}

\item{prob_CN_cnloh_interstitial_length}{...}

\item{prob_CN_cnloh_terminal_length}{...}

\item{rate_driver}{...}

\item{rate_passenger}{...}

\item{bound_driver}{...}

\item{bound_average_ploidy}{...}

\item{bound_homozygosity}{...}

\item{SFS_totalsteps}{...}

\item{prob_coverage}{...}

\item{alpha_coverage}{...}

\item{lower_limit_cell_counts}{...}

\item{lower_limit_alt_counts}{...}

\item{lower_limit_tot_counts}{...}

\item{gc}{...}

\item{gc_slope}{...}

\item{gc_int}{...}

\item{sigma1}{...}

\item{num_reads}{...}
}
\description{
Produce simulations and outputs files as requested by user
}
