% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUILD_initial_population.r
\name{BUILD_initial_population}
\alias{BUILD_initial_population}
\title{Update the initial population for clones}
\usage{
BUILD_initial_population(
  model_variables = list(),
  cell_count = 1,
  CN_matrix = list(),
  drivers = list()
)
}
\arguments{
\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}

\item{cell_count}{An integer value specifying the initial population's cell count. Default is 1.}

\item{CN_matrix}{A dataframe specifying the CN profile of the initial cell population.}

\item{drivers}{??? A list specifying the driver genes the initial population of cells have.}
}
\description{
\code{BUILD_initial_population} returns an updated object specifying the characteristics, including any drivers and CN profile, of the initial population of cell(s).
}
\examples{

model_variables_base <- BUILD_initial_population(
   model_variables = model_variables_base,
   cell_count = cell_count,
   CN_matrix = CN_matrix,
   drivers = drivers)

}
