% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_bulk_arm_CN.r
\name{bulk_arm_CN_assign_paras}
\alias{bulk_arm_CN_assign_paras}
\title{Assign parameters to proper positions}
\usage{
bulk_arm_CN_assign_paras(model_variables, parameter_IDs, parameters)
}
\arguments{
\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}

\item{parameter_IDs}{???}

\item{parameters}{???}
}
\description{
\code{bulk_arm_CN_assign_paras} returns an updated object that has parameters assigned to the correct position???
}
\examples{
bulk_arm_CN_assign_paras(model_variables, parameter_IDs, parameters)

}
