% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECK_model_variables.r
\name{CHECK_model_variables}
\alias{CHECK_model_variables}
\title{Check/Clean up the model variables}
\usage{
CHECK_model_variables(model_variables)
}
\arguments{
\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}
}
\description{
\code{CHECK_model_variable} returns an updated object that checks the specified model variables to clean up the libraries of CN information and chromosome arm selection rates.
}
\examples{

CHECK_model_variables(model_variables)


}
