% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_bulk_arm_CN.r
\name{fitting_bulk_arm_CN}
\alias{fitting_bulk_arm_CN}
\title{ABC fitting for the chromosome arms}
\usage{
fitting_bulk_arm_CN(
  library_name,
  model_name,
  model_variables,
  copynumber_DATA,
  type_sample_DATA = "individual",
  type_cn_DATA = "bin",
  list_parameters,
  list_parameters_library,
  list_targets,
  list_targets_library,
  bound_freq = 0.1,
  ntree = 200,
  library_shuffle = FALSE,
  n_cores = NULL,
  R_libPaths = NULL,
  folder_workplace = NULL
)
}
\arguments{
\item{library_name}{A string to be used in the resulting rda's name.}

\item{model_name}{A string to be used to identify the model. ???}

\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}

\item{copynumber_DATA}{A dataframe??? containing the copynumber profiles of the data for the parameters to be fit against.}

\item{type_sample_DATA}{???}

\item{type_cn_DATA}{???}

\item{list_parameters}{???}

\item{list_parameters_library}{A dataframe consisting of the parameters to be fitted. The dataframe should have columns "Variable", "Type", "Lower_bound", "Upper_bound".}

\item{list_targets}{A list specifying the target parameter(s) to be fitted by ABC.}

\item{list_targets_library}{???}

\item{bound_freq}{??? Default is 0.1.}

\item{ntree}{An integer specifying the number of trees to grow in the forest. Default is 200 trees}

\item{library_shuffle}{???}

\item{n_cores}{An integer number specifying the number of cores to be used for parallel processing. ???}

\item{R_libPaths}{??? The location of the necessary R packages to run CINner when using an HPC cluster. Default is NULL.}

\item{folder_workplace}{The location of the folder to contain the output files and plots. Default is NULL.}

\item{ABC_simcount}{An integer number specifying the number of simulations for a given parameters set. Default is 10000.}
}
\description{
\code{fitting_bulk_arm_CN} returns the results of ABCrf fitting for the specified parameters using the simulated library. It increases the size of the simulated library via bootstrapping, and it saves the results of the parameter fitting to an output rda file. The function also returns a plot of the prior, posterior, and chosen best parameter value for the fitted parameters.
}
\examples{
???
fitting_bulk_arm_CN(
       library_name = library_name,
       model_name = cancer_type,
       model_variables = model_variables,
       copynumber_DATA = copynumber_DATA,
       list_parameters = list_parameters,
       list_parameters_library = list_parameters_library,
       list_targets = list_targets,
       list_targets_library = list_targets_library,
       bound_freq = 0.1,
       library_shuffle = TRUE,
       folder_workplace = library_name,
       R_libPaths = R_libPaths)

}
