% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_bulk_arm_CN.r
\name{library_bulk_arm_CN}
\alias{library_bulk_arm_CN}
\title{Build a library of simulated arm copy number profiles}
\usage{
library_bulk_arm_CN(
  library_name,
  model_variables,
  list_parameters,
  list_targets,
  ABC_simcount = 10000,
  n_cores = NULL,
  n_samples = 100,
  R_libPaths = NULL
)
}
\arguments{
\item{library_name}{A string to be used in the resulting rda's name.}

\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}

\item{list_parameters}{A list of the parameter IDs to be fitted. ???}

\item{list_targets}{A list specifying the target parameter(s) to be fitted by ABC.}

\item{ABC_simcount}{An integer number specifying the number of simulations for a given parameters set. Default is 10000.}

\item{n_cores}{An integer number specifying the number of cores to be used for parallel processing. ???}

\item{n_samples}{An integer value specifying the number of values from the prior distribution that are randomly selected for each simulation. Default is 100.}

\item{R_libPaths}{??? The location of the necessary R packages to run CINner when using an HPC cluster. Default is NULL.}
}
\description{
\code{library_bulk_arm_CN} returns an rda object that contains the results of the simulated clonal evolution as well as the statistics about the clones from the resulted sampling.
}
\examples{


library_bulk_arm_CN(
   library_name = library_name,
   model_variables = model_variables,
   list_parameters = list_parameters_library,
   list_targets = list_targets_library,
   ABC_simcount = 2000,
   n_samples = 100,
   R_libPaths = R_libPaths)

}
