% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_bulk_gene.r
\name{library_bulk_gene}
\alias{library_bulk_gene}
\title{Fitting driver gene data for a cancer type from PCAWG}
\usage{
library_bulk_gene(
  library_name,
  model_variables,
  list_parameters,
  list_targets,
  ABC_simcount = 10000,
  n_cores = NULL,
  n_samples = 100,
  R_libPaths = NULL
)
}
\arguments{
\item{library_name}{A string to be used in the resulting rda's name.}

\item{model_variables}{??? The object specifying the values of the parameters the simulation uses.}

\item{list_parameters}{A list of the parameter IDs to be fitted. ???}

\item{list_targets}{A list specifying the target parameter(s) to be fitted by ABC.}

\item{ABC_simcount}{An integer number specifying the number of simulations for a given parameters set. Default is 10000.}

\item{n_cores}{An integer number specifying the number of cores to be used for parallel processing. ???}

\item{n_samples}{An integer value specifying the number of values from the prior distribution that are randomly selected for each simulation. Default is 100.}

\item{R_libPaths}{??? The location of the necessary R packages to run CINner when using an HPC cluster. Default is NULL.}
}
\description{
Fitting driver gene data for a cancer type from PCAWG
}
\examples{

list_parameters <- data.frame(matrix(ncol = 4, nrow = 0))
colnames(list_parameters) <- c("Variable", "Type", "Lower_bound", "Upper_bound")
list_parameters[nrow(list_parameters) + 1, ] <- c("prob_CN_focal_deletion", "CNA_probability", 1e-6, 1e-5)

list_targets <- model_variables$driver_library$Gene_ID


}
