% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_bulk_arm_CN.r
\name{statistics_bulk_arm_WGD_status}
\alias{statistics_bulk_arm_WGD_status}
\title{Find statistics from PCAWG data and plot the relevant analyses}
\usage{
statistics_bulk_arm_WGD_status(
  plotname,
  DATA_cancer_types,
  DATA_cancer_type_sample_ids,
  DATA_cancer_type_cn,
  DATA_wgd,
  model_variables
)
}
\arguments{
\item{plotname}{A string value to be incorporated into the output plots.}

\item{DATA_cancer_types}{A list containing the names of the cancer types from the data that were fitted.}

\item{DATA_cancer_type_sample_ids}{A list containing the sample IDs from the data.}

\item{DATA_cancer_type_cn}{??? A list containing the copynumber data for each sample from the data.}

\item{DATA_wgd}{A list containing the sample IDs that are considered WGD.}

\item{model_variables}{The model variables used to specify the parameter values used for the simulation.}
}
\description{
\code{statistics_bulk_arm_WGD_status} finds the statistics from the given data and plots the relevant analyses.
}
\examples{

copynumber_PCAWG <- read_excel(system.file("pcawg_specimen_histology_August2016_v9.xlsx", package = "CINner"))
PCAWG_cancer_types <- unique(copynumber_PCAWG$histology_abbreviation)
PCAWG_cancer_types <- PCAWG_cancer_types[which(!is.na(PCAWG_cancer_types))

PCAWG_cancer_type_sample_ids <- vector("list", length(PCAWG_cancer_types))
for (i in 1:length(PCAWG_cancer_types)) {
   sample_ids <- copynumber_PCAWG$tcga_sample_uuid[which(copynumber_PCAWG$histology_abbreviation == PCAWG_cancer_types[i])]
   for (j in 1:length(sample_ids)) {
       sample_id <- sample_ids[j]
       if (!file.exists(system.file(paste0("consensus.20170119.somatic.cna.annotated/", sample_id, ".consensus.20170119.somatic.cna.annotated.txt"), package = "CINner"))) next
       tmp <- read.table(system.file(paste0("consensus.20170119.somatic.cna.annotated/", sample_id, ".consensus.20170119.somatic.cna.annotated.txt"), package = "CINner"), header = TRUE)
       if (max(tmp$star, na.rm = TRUE) < 2) next
       PCAWG_cancer_type_sample_ids[[i]] <- c(PCAWG_cancer_type_sample_ids[[i]], sample_id)
   }
}

copynumber_DATA_cancer_types <- vector("list", length(PCAWG_cancer_types))
for (i in 1:length(PCAWG_cancer_types)) {
   cancer_type <- PCAWG_cancer_types[i]
   cancer_type_sample_ids <- PCAWG_cancer_type_nonWGD_sample_ids[[i]]
   copynumber_DATA_ls <- vector("list", length(cancer_type_sample_ids))
   for (j in 1:length(cancer_type_sample_ids)) {
       sample_id <- cancer_type_sample_ids[j]
       tmp <- read.table(system.file(paste0("consensus.20170119.somatic.cna.annotated/", sample_id, ".consensus.20170119.somatic.cna.annotated.txt"), package = "CINner"), header = TRUE)
       if (any(is.na(tmp$star))) tmp <- tmp[-which(is.na(tmp))]
       tmp <- tmp[which(tmp$star >= 2), ]
       tmp$donor_unique_id <- sample_id
       copynumber_DATA_ls[[j]] <- tmp
   }
   copynumber_DATA_cancer_types[[i]] <- rbindlist(copynumber_DATA_ls)
}

PCAWG_wgd <- read.table(system.file("consensus.20170218.purity.ploidy.txt", package = "CINner"), header = TRUE)

statistics_bulk_arm_WGD_status(
   plotname = "ALL_PCAWG",
   DATA_cancer_types = PCAWG_cancer_types,
   DATA_cancer_type_sample_ids = PCAWG_cancer_type_sample_ids,
   DATA_cancer_type_cn = copynumber_DATA_cancer_types,
   DATA_wgd = PCAWG_wgd,
   model_variables = model_variables)

}
