% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULATOR_VARIABLES_for_simulation.r
\name{SIMULATOR_VARIABLES_for_simulation}
\alias{SIMULATOR_VARIABLES_for_simulation}
\title{Add together two numbers}
\usage{
SIMULATOR_VARIABLES_for_simulation(model)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
