% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULATOR_VARIABLES_for_simulation.r
\name{SIMULATOR_VARIABLES_for_simulation}
\alias{SIMULATOR_VARIABLES_for_simulation}
\title{Input variables to simulate a cancer model}
\usage{
SIMULATOR_VARIABLES_for_simulation(model)
}
\arguments{
\item{model}{Name of model.}
}
\value{
Nothing.
}
\description{
Input variables to simulate a cancer model
}
\examples{
SIMULATOR_VARIABLES_for_simulation("FALLOPIAN-TUBES-NEUTRAL")
}
